% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxgrob.R
\name{add_box}
\alias{add_box}
\title{Add node}
\usage{
add_box(prev_box = NULL, txt, dist = 0.02)
}
\arguments{
\item{prev_box}{Previous node object, the created new node will be vertically
aligned with this node. Left this as `NULL` if this is the first node. The first
node will be aligned in the top center.}

\item{txt}{Text in the node. If the `prev_box` is a horizontally aligned multiple
nodes, a vector of with the same length must be provided.}

\item{dist}{Distance between previous node, including the distance between the
side node.}
}
\description{
Create vertically aligned labeled box for the nodes.
}
\examples{
\dontrun{
node1 <- add_box(txt = "Screened (n=300)")
}

}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}}
}
