% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_hori.R
\name{align_hori}
\alias{align_hori}
\title{Align list of box grobs horizontally}
\usage{
align_hori(boxlist)
}
\arguments{
\item{boxlist}{A list of box grobs.}
}
\value{
A list of box grobs
}
\description{
This function will align all box with the lowest.
}
\examples{

}
\keyword{internal}
