% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxgrob.R
\name{box_text}
\alias{box_text}
\title{Generate label and bullet points}
\usage{
box_text(x, label, sum_only = TRUE)
}
\arguments{
\item{x}{A list or a vector to be used.}

\item{label}{A character string as a label at the beginning of the text label.}

\item{sum_only}{If shows total only (default). If the value is `FALSE`, the
values will be tabulated with bullet points.}
}
\value{
A character string of vector.
}
\description{
This function use the data to generate label and bullet points for the box.
}
\examples{
val <- data.frame(am = factor(ifelse(mtcars$am == 1, "Automatic", "Manual")),
                 car = row.names(mtcars))
 
 box_text(val$car, label = "Cars in the data")
 box_text(val$car, label = "Cars in the data", sum_only = FALSE)
 box_text(split(val$car, val$am), label = "Cars in the data")
 box_text(split(val$car, val$am), label = "Cars in the data", sum_only = FALSE)


}
