% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{build_consort}
\alias{build_consort}
\title{Build consort diagram}
\usage{
build_consort(consort_list, label_list = NULL)
}
\arguments{
\item{consort_list}{A list of nodes.}

\item{label_list}{A list of label nodes.}
}
\description{
This function has been deprecated, no need to use this function.
Please check the vignettes
}
