% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\docType{package}
\name{consort}
\alias{consort-package}
\alias{consort}
\title{Create Consort diagram}
\description{
To make it easy to create CONSORT diagrams for the transparent reporting
 of participant allocation in randomized, controlled clinical trials. This
  is done by creating a standardized disposition data, and using this data
   as the source for the creation a standard CONSORT diagram. Human effort
    by supplying text labels on the node can also be achieved.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adayim/consort/}
  \item Report bugs at \url{https://github.com/adayim/consort/issues}
}

}
\author{
\strong{Maintainer}: Alim Dayim \email{ad938@cam.ac.uk} (\href{https://orcid.org/0000-0001-9998-7463}{ORCID})

}
\keyword{internal}
