% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{grid.draw.consort.plot}
\alias{grid.draw.consort.plot}
\title{grid.draw method}
\usage{
\method{grid.draw}{consort.plot}(x)
}
\arguments{
\item{x}{A consort.plot object.}
}
\description{
This method is to support saving plots with `ggplot2::ggsave`.
}
