% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing-strings}
\alias{missing-strings}
\alias{\%??\%}
\title{Empty string if missing}
\usage{
x \%??\% y
}
\arguments{
\item{x, y}{If \code{x} is missing, returns ""; otherwise returns \code{y} or \code{x} if \code{y}
is \code{NULL}.}
}
\description{
If the LHS is missing, replaces with "". If the LHS is not
missing, replaces with the RHS (or the LHS if RHS is \code{NULL}).
}
\examples{
NULL \%??\% "other"
"apple" \%??\% "other"
"apple" \%??\% NULL
}
\seealso{
Other utilities: \code{\link{null-default}}
}
\concept{utilities}
