% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\title{Null or Default}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If \code{x} is NULL, will return \code{y}; otherwise returns \code{x}.}
}
\description{
Null or default infix function borrowed from \link{purrr}.
}
\examples{
1 \%||\% 2
NULL \%||\% 2
}
\seealso{
Other utilities: \code{\link{missing-strings}}
}
\concept{utilities}
