% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.consort}
\alias{plot.consort}
\alias{print.consort}
\title{Add methods to print function}
\usage{
\method{plot}{consort}(x, grViz = FALSE, ...)

\method{print}{consort}(x, grViz = FALSE, ...)
}
\arguments{
\item{x}{A \code{consort} object.}

\item{grViz}{If use \link[DiagrammeR]{grViz} to print the plot. 
Default is \code{FALSE} to use \link[grid]{grid.draw}}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Method for plot objects and display the output in on a grid device.
}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}},
\code{\link{add_side_box}}, \link[grid]{grid.draw}
}
