% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.consort.plot}
\alias{print.consort.plot}
\title{Add methods to print function}
\usage{
\method{print}{consort.plot}(x, ...)
}
\arguments{
\item{x}{A \code{consort.list}, \code{consort.plot} or \code{consort}.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Method for create viewports for the \code{consort.list}, \code{consort.plot} 
or \code{consort} objects and display the output in on a grid device.
}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}},
\code{\link{add_side_box}}, \code{\link{consort_plot}},\code{\link{build_consort}}
}
