% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.consort}
\alias{print.consort}
\alias{print.consort.list}
\alias{grid.draw.consort.plot}
\title{Print Consort Plots}
\usage{
\method{print}{consort.list}(x, ...)

\method{print}{consort}(x, ...)

\method{grid.draw}{consort.plot}(x)
}
\arguments{
\item{x}{A consort.plot object.}

\item{...}{further arguments passed to \code{print} or \code{grid::grid.draw}.}
}
\value{
None
}
\description{
This method is to support saving plots with `ggplot2::ggsave`.
}
