% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoRt-package.R
\docType{data}
\name{study_data}
\alias{study_data}
\title{Example Study Data}
\format{A data frame with 500 rows and 12 variables:
\describe{
\item{\code{id}}{integer. Participant ID}
\item{\code{metadata}}{character. Name of participant, representative of
metadata associated with the participant but unrelated to trial design.}
\item{\code{eligible}}{character. Eligibility status of the participant, where
\code{"Eligible"} indicates the participant was eligible and any other value
describes the ineligibility group.}
\item{\code{allocation}}{character. Allocation of the trial participant into
the \code{"Intervention"} or \code{"Placebo"} group.}
\item{\code{intervention}}{logical. Did the participant receive the
intervention? \code{TRUE} or \code{FALSE} if the participant was allocated to the
intervention group, \code{NA} otherwise.}
\item{\code{intervention_excl_reason}}{character. Reasons describing the
circumstances leading to the exclusion of the participant from the
intervention group, \code{NA} if not excluded or in the intervention group.}
\item{\code{placebo}}{logical. Did the participant receive the placebo? \code{TRUE}
or \code{FALSE} if the participant was allocated to the placebo group, \code{NA}
otherwise.}
\item{\code{placebo_excl_reason}}{character. Reasons describing the
circumstances leading to the exclusion of the participant from the
placebo group, \code{NA} if not excluded or in the intervention group.}
\item{\code{lost_reason}}{character. Reasons describing the circumstances by
which the participant was lost to follow-up, or \code{NA} if the participant
did not receive the intervention or placebo.}
\item{\code{discontinued_reason}}{character. Reasons describing the
circumstances by which the participant discontinued their participant,
or \code{NA} if the participant did not receive the intervention or placebo.}
\item{\code{analyzed}}{logical. Was the participant's data analyzed? \code{NA} if
the participant was excluded prior to this stage.}
\item{\code{analyzed_excl_reason}}{character. Reasons describing the exclusion
of a participant from analysis if they otherwise reached this stage of the
trial.}
}}
\usage{
study_data
}
\description{
A dataset of 500 study participants with columns tracking their participation
in a randomized trial.
}
\keyword{datasets}
