% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_latex}
\alias{write_latex}
\title{Write LaTeX File and Compile to PDF or Image}
\usage{
write_latex(x, path, ..., density = 72)
}
\arguments{
\item{x}{LaTeX document as character string}

\item{path}{<chr> File name for drawing output}

\item{density}{resolution to render raster image output via
\link[magick:image_write]{magick::image_write}}
}
\description{
Write LaTeX File and Compile to PDF or Image
}
