% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxgrob.R
\name{add_label_box}
\alias{add_label_box}
\title{Add a vertically aligned label nodes on the left side.}
\usage{
add_label_box(ref_box, txt)
}
\arguments{
\item{ref_box}{Reference node to which the label will be horizontally aligned.}

\item{txt}{Text in the node.}
}
\description{
In a consort diagram, this can be used to indicate different stage.
}
\examples{
\dontrun{
lab1 <- add_label_box(node1, txt = "Screening")
lab2 <- add_label_box(node4, txt = "Randomized")
}
}
