% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxgrob.R
\name{add_side_box}
\alias{add_side_box}
\title{Add side node}
\usage{
add_side_box(prev_box, txt, dist = 0.02)
}
\arguments{
\item{prev_box}{Previous node object, the created new node will be aligned
at the right bottom of the `prev_box`.}

\item{txt}{Text in the node. If the `prev_box` is a horizontally aligned multiple
nodes, a vector of with the same length must be provided.}

\item{dist}{Distance between previous node, including the distance between the
side node.}
}
\description{
Add an exclusion node on the right side.
}
\examples{
\dontrun{
txt1_side <- "Excluded (n=11):\n\u2022 MRI not collected (n=3)\n\u2022 Other (n=8)"
node3 <- add_side_box(node1, txt = txt1_side)
}
}
\seealso{
\code{\link{add_box}},\code{\link{add_split}}
}
