% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxgrob.R
\name{build_consort}
\alias{build_consort}
\title{Build consort diagram}
\usage{
build_consort(consort_list, label_list = NULL)
}
\arguments{
\item{consort_list}{A list of nodes.}

\item{label_list}{A list of label nodes.}
}
\description{
Build consort diagram
}
\examples{
\dontrun{
build_consort(consort_list = list(node1, node3, node4, node1_sp,
                                   side1_sp, node2_sp),
              label_list   = list(lab1, lab2, lab3))

}


}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}},
\code{\link{add_side_box}}
}
