% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grviz.R
\name{build_grviz}
\alias{build_grviz}
\title{Build consort diagram}
\usage{
build_grviz(x)
}
\arguments{
\item{x}{A conosrt object.}
}
\value{
A \code{Graphviz} code
}
\description{
Build a \code{grob} consort diagram, use this if you want
to save plots with \code{\link[ggplot2]{ggsave}}
}
\examples{
txt1 <- "Population (n=300)"
txt1_side <- "Excluded (n=15): \n
              \u2022 MRI not collected (n=3)\n
              \u2022 Tissues not collected (n=4)\n
              \u2022 Other (n=8)"

g <- add_box(txt = txt1)

g <- add_side_box(g, txt = txt1_side)

g <- add_box(g, txt = "Randomized (n=200)")
plot(g, grViz = TRUE)

}
\seealso{
\code{\link[DiagrammeR]{grViz}}
}
