% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\docType{package}
\name{consort-package}
\alias{consort-package}
\title{Create consort diagram}
\description{
to make it easy to create CONSORT diagrams for the transparent reporting
 of participant allocation in randomized, controlled clinical trials. This
  is done by creating a standardized disposition data, and using this data
   as the source for the creation a standard CONSORT diagram. Human effort
    by supplying text labels on the node can also be achieved.
}
