% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoRt-package.R
\docType{package}
\name{consoRt-package}
\alias{consoRt}
\alias{consoRt-package}
\title{consoRt: Create CONSORT Diagrams with tikz}
\description{
Create CONSORT diagrams with tikz in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GerkeLab/consoRt}
  \item Report bugs at \url{https://github.com/GerkeLab/consoRt/issues}
}

}
\author{
\strong{Maintainer}: Travis Gerke \email{Travis.Gerke@moffitt.org}

Authors:
\itemize{
  \item Jordan Creed \email{Jordan.H.Creed@moffitt.org}
  \item Garrick Aden-Buie \email{Garrick.Aden-Buie@moffitt.org}
}

}
\keyword{internal}
