% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_box.R
\name{.add_box}
\alias{.add_box}
\title{Create node vertically align with the previous one}
\usage{
.add_box(prev_vert, prev_side = NULL, txt, just = "center", dist = 0.02, ...)
}
\arguments{
\item{txt}{Text in the node. If the `prev_box` is a horizontally aligned multiple
nodes, a vector of with the same length must be provided.}

\item{just}{The justification for the text: left, center or right.}

\item{dist}{Distance between previous node, including the distance between the
side node.}

\item{...}{Other parameters pass to \link{textbox},}
}
\description{
Create node vertically align with the previous one
}
\keyword{internal}
