% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords}
\alias{get_coords}
\title{Get the coordinates of the box object}
\usage{
get_coords(x)
}
\arguments{
\item{x}{A box object}
}
\value{
A list of coordinates will return:
\item{left}{Left (x-min) side coordinate.}
\item{right}{Right (x-max) side coordinate.}
\item{bottom}{Bottom (y-min) side coordinate.}
\item{top}{Top (y-max) side coordinate.}
\item{x}{X (center x) coordinate.}
\item{y}{Y (center y) coordinate.}
\item{width}{Width of the box.}
\item{height}{Height of the box.}
\item{half_width}{Half width of the box.}
\item{half_height}{Half height of the box.}
}
\description{
This function will get the coordinates of the box object.
}
\examples{
fg <- textbox(text = "This is a test")
get_coords(fg)

}
