% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.consort}
\alias{plot.consort}
\title{Add methods to print function}
\usage{
\method{plot}{consort}(x, ...)
}
\arguments{
\item{x}{A \code{consort} object.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Method for plot objects and display the output in on a grid device.
}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}},
\code{\link{add_side_box}}, \link[grid]{grid.draw}
}
