% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.consort.plot}
\alias{print.consort.plot}
\title{Add methods to print function}
\usage{
\method{print}{consort.plot}(x, ...)
}
\arguments{
\item{x}{A consort.list, consort.plot or consort object.}

\item{...}{Not used.}
}
\description{
Add methods to print function
}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}},
\code{\link{add_side_box}}, \code{\link{consort_plot}},\code{\link{build_consort}}
}
